backend/requirements.txt
Django==5.0.10
djangorestframework==3.15.2
psycopg2-binary==2.9.11
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
celery==5.4.0
redis==5.0.8
django-filter==24.3
django-cors-headers==4.4.0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
pytest==8.3.2
pytest-django==4.9.0
